Move template to one file
