--- utils/templates/tl1_preprocess.h
@@ -1,7 +1,0 @@
-template<int K>
-void preprocessor_k(void* B, void* LUT_Scales, void* QLUT) {
-  partial_max_reset((&(((bitnet_float_type*)LUT_Scales)[0])));
-  per_tensor_quant(K, (&(((bitnet_float_type*)LUT_Scales)[0])), (&(((bitnet_float_type*)B)[0])));
-  
-  lut_ctor<K>((&(((int8_t*)QLUT)[0])), (&(((bitnet_float_type*)B)[0])), (&(((bitnet_float_type*)LUT_Scales)[0])));
-}
